requirements.txt
requests==2.32.5
psycopg2-binary==2.9.11
asyncpg==0.29.0
apache-airflow-providers-celery>=3.3.0
apache-airflow-providers-fab>=1.0.0
celery==5.5.3